Refactor common back in, flat directory structure makes no sense
